[refactor] User 서비스 DDD 구조 적용
