+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/domain/model/Role.java
@@ -1,4 +1,4 @@
-package com.sparta_logistics.auth.Entity;
+package com.sparta_logistics.auth.domain.model;
 
 import lombok.Getter;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
