+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/infrastructure/config/AuthConfig.java
@@ -1,8 +1,10 @@
-package com.sparta_logistics.auth.Security;
+package com.sparta_logistics.auth.infrastructure.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta_logistics.auth.jwt.JwtAuthenticationFilter;
-import com.sparta_logistics.auth.jwt.JwtAuthorizationFilter;
+import com.sparta_logistics.auth.application.Security.JwtUtil;
+import com.sparta_logistics.auth.application.Security.UserDetailsServiceImpl;
+import com.sparta_logistics.auth.presentation.filter.JwtAuthenticationFilter;
+import com.sparta_logistics.auth.presentation.filter.JwtAuthorizationFilter;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
