+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/infrastructure/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.sparta_logistics.auth.Security;
+package com.sparta_logistics.auth.infrastructure.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
