+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/presentation/Dto/DeletedUserInfoResponseDto.java
@@ -1,7 +1,7 @@
-package com.sparta_logistics.auth.Dto;
+package com.sparta_logistics.auth.presentation.Dto;
 
-import com.sparta_logistics.auth.Entity.Role;
-import com.sparta_logistics.auth.Entity.User;
+import com.sparta_logistics.auth.domain.model.Role;
+import com.sparta_logistics.auth.domain.model.User;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Getter;
@@ -11,8 +11,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class UserInfoResponseDto {
-
+public class DeletedUserInfoResponseDto
+{
   private UUID userId;
   private String userName;
   private String slackId;
@@ -23,7 +23,11 @@ public class UserInfoResponseDto {
   private String createdBy;
   private String updatedBy;
 
-  public UserInfoResponseDto(User user){
+  private boolean isDeleted ;
+  private LocalDateTime deletedAt;
+  private String deletedBy;
+
+  public DeletedUserInfoResponseDto(User user){
     this.userId = user.getUserId();
     this.userName = user.getUserName();
     this.slackId = user.getSlackId();
@@ -32,6 +36,9 @@ public class UserInfoResponseDto {
     this.updatedAt = user.getUpdatedAt();
     this.createdBy = user.getCreatedBy();
     this.updatedBy = user.getUpdatedBy();
+    this.isDeleted = user.isDeleted();
+    this.deletedAt = user.getDeletedAt();
+    this.deletedBy = user.getDeletedBy();
   }
 
 }
