+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/presentation/Dto/SignInRequestDto.java
@@ -1,16 +1,14 @@
-package com.sparta_logistics.auth.Dto;
+package com.sparta_logistics.auth.presentation.Dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.sparta_logistics.auth.Entity.Role;
 import lombok.Getter;
 
 @Getter
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class UserInfoResDto {
+public class SignInRequestDto {
 
   private String userName;
-  private String slackId;
-  private Role role;
+  private String password;
 
 }
