+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/presentation/Dto/SignUpRequestDto.java
@@ -1,8 +1,8 @@
-package com.sparta_logistics.auth.Dto;
+package com.sparta_logistics.auth.presentation.Dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.sparta_logistics.auth.Entity.Role;
+import com.sparta_logistics.auth.domain.model.Role;
 import lombok.Getter;
 import jakarta.validation.constraints.Pattern;
