+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/presentation/Dto/UserInfoResDto.java
@@ -1,14 +1,16 @@
-package com.sparta_logistics.auth.Dto;
+package com.sparta_logistics.auth.presentation.Dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import com.sparta_logistics.auth.domain.model.Role;
 import lombok.Getter;
 
 @Getter
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class SignInRequestDto {
+public class UserInfoResDto {
 
   private String userName;
-  private String password;
+  private String slackId;
+  private Role role;
 
 }
