+++ com.sparta_logistics.auth/src/main/java/com/sparta_logistics/auth/presentation/Dto/UserUpdateRequestDto.java
@@ -1,13 +1,9 @@
-package com.sparta_logistics.auth.Dto;
+package com.sparta_logistics.auth.presentation.Dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.sparta_logistics.auth.Entity.Role;
-import com.sparta_logistics.auth.Entity.User;
-import jakarta.validation.constraints.Pattern;
-import lombok.Builder;
+import com.sparta_logistics.auth.domain.model.Role;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
